Change Table view, add message and put common components together
