+++ front_end/Routing.tsx
@@ -10,8 +10,8 @@ import Login from './components/auth/Login'
 import Registration from './components/auth/Registration'
 import GlobalMessage from './components/GlobalMessage'
 import Header from './components/Header'
-import ExpenditureLogsTable from './components/expenditure/ExpenditureLogsTable'
-import IncomeLogsTable from './components/income/IncomeLogsTable'
+import ExpenditureLogsTable from './components/logs/expenditure/ExpenditureLogsTable'
+import IncomeLogsTable from './components/logs/income/IncomeLogsTable'
 import MonthlyExpenditureTable from './components/monthly/MonthlyExpendiitureTable'
 import AssetPage from './components/asset/AssetPage'
