+++ front_end/components/Hello.tsx
@@ -3,8 +3,8 @@ import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import Cancellation from './auth/Cancellation'
 import AssetPage from './asset/AssetPage'
-import ExpenditureLogsTable from './expenditure/ExpenditureLogsTable'
-import IncomeLogsTable from './income/IncomeLogsTable'
+import ExpenditureLogsTable from './logs/expenditure/ExpenditureLogsTable'
+import IncomeLogsTable from './logs/income/IncomeLogsTable'
 import MonthlyExpenditureTable from './monthly/MonthlyExpendiitureTable'
 
 interface HelloProps {}
