+++ front_end/components/logs/expenditure/EnhancedTableHead.tsx
