+++ front_end/components/logs/income/IncomeLogsTable.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 import * as moment from 'moment';
 import { useSelector, useDispatch } from 'react-redux'
 
-import ExpenditureLog from '../../models/ExpenditureLog';
-import EdtingExpenditureLog from './EditingExpenditureLog';
-import { deleteExpenditureLog } from '../../services/ExpenditureLogService';
-import { editExpenditureLog } from '../../modules/ExpenditureLogModule';
+import IncomeLog from '../../../models/IncomeLog';
+import EdtingIncomeLog from './EditingIncomeLog';
+import { editIncomeLog } from '../../../modules/IncomeLogModule';
 import EnhancedTableHead from './EnhancedTableHead'
 import EnhancedTableToolbar from './EnhancedTableToolbar'
+import { deleteIncomeLog } from '../../../services/IncomeLogService';
 
 import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
 import Table from '@material-ui/core/Table';
@@ -18,18 +18,19 @@ import TablePagination from '@material-ui/core/TablePagination';
 import TableRow from '@material-ui/core/TableRow';
 import Paper from '@material-ui/core/Paper';
 import Checkbox from '@material-ui/core/Checkbox';
-import IconButton from '@material-ui/core/IconButton';
-import Tooltip from '@material-ui/core/Tooltip';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Switch from '@material-ui/core/Switch';
-import DeleteForeverIcon from '@material-ui/icons/DeleteForever'
-import CreateExpenditureLogModal from './CreateExpenditureLogModal';
+import CreateIncomeLogModal from './CreateIncomeLogModal';
+import LoadingIcon from '../../LoadingIcon';
+import DeleteAlert from '../common/DeleteAlert';
+import { successMessage, succesmMessages } from '../../GlobalMessage';
+
 
 interface tableData {
   title: string;
   amount: number;
   content: string;
-  paidAt: Date;
+  earnedAt: Date;
   edit: string;
   delete: string;
 }
@@ -71,7 +72,8 @@ const useStyles = makeStyles((theme: Theme) =>
       maxHeight: 671
     },
     root: {
-      width: '70%',
+      width: '90%',
+      margin: '0 auto'
     },
     paper: {
       width: '100%',
@@ -94,36 +96,36 @@ const useStyles = makeStyles((theme: Theme) =>
   }),
 );
 
-const ExpenditureLogsTable: React.FC = () => {
+const IncomeLogsTable: React.FC = () => {
   const { useState } = React
   const [order, setOrder] = useState<Order>('asc')
   const [orderBy, setOrderBy] = useState<keyof tableData>('amount')
   const [page, setPage] = useState(0)
   const [dense, setDense] = useState(true)
   const [rowsPerPage, setRowsPerPage] = useState<number>(1)
-  const [checkedLogs, setCheckedLogs] = useState<ExpenditureLog[]>([])
-  const [rows, setRows] = useState<ExpenditureLog[]>([])
+  const [checkedLogs, setCheckedLogs] = useState<IncomeLog[]>([])
+  const [rows, setRows] = useState<IncomeLog[]>([])
 
   const classes = useStyles()
 
-  const expenditureLogs = useSelector(state => state.expenditureLogs)
+  const incomeLogs = useSelector(state => state.incomeLogs)
   const dispatch = useDispatch()
 
   React.useEffect(() => {
-    setRowsPerPage(expenditureLogs.length)
+    setRowsPerPage(incomeLogs.length)
 
-    const logs = expenditureLogs.map((log: ExpenditureLog) => {
-      return new ExpenditureLog(
-        log.title,
-        log.amount,
-        log.content,
-        log.paidAt,
-        log.id
+    const logs = incomeLogs.map((incomeLog: IncomeLog) => {
+      return new IncomeLog(
+        incomeLog.title,
+        incomeLog.amount,
+        incomeLog.content,
+        incomeLog.earnedAt,
+        incomeLog.id
       )
     })
 
     setRows(logs)
-  }, [expenditureLogs])
+  }, [incomeLogs])
 
   const handleRequestSort = (event: React.MouseEvent<unknown>, property: keyof tableData) => {
     const isAsc = orderBy === property && order === 'asc'
@@ -139,35 +141,35 @@ const ExpenditureLogsTable: React.FC = () => {
     setCheckedLogs([])
   }
 
-  const removeCheck = (expenditureLog: ExpenditureLog) => {
+  const removeCheck = (incomeLog: IncomeLog) => {
     setCheckedLogs(checkedLogs.filter(log => {
-      if (log.id !== expenditureLog.id) return log
+      if (log.id !== incomeLog.id) return log
     }))
   }
 
-  const handleCheckClick = (event: React.MouseEvent<unknown>, expenditureLog: ExpenditureLog) => {
+  const handleCheckClick = (event: React.MouseEvent<unknown>, incomeLog: IncomeLog) => {
     const logIds = checkedLogs.map(log => { return log.id})
 
-    const selectedIndex = logIds.indexOf(expenditureLog.id)
+    const selectedIndex = logIds.indexOf(incomeLog.id)
     let newSelected: number[] = []
 
     if (selectedIndex === -1) { // Adding: 存在しない場合
-      newSelected = newSelected.concat(logIds, expenditureLog.id)
+      newSelected = newSelected.concat(logIds, incomeLog.id)
 
-      setCheckedLogs(checkedLogs.concat(expenditureLog))
+      setCheckedLogs(checkedLogs.concat(incomeLog))
     } else if (selectedIndex === 0) { // Removing: 最初に存在する場合
       newSelected = newSelected.concat(logIds.slice(1))
-      removeCheck(expenditureLog)
+      removeCheck(incomeLog)
     } else if (selectedIndex === logIds.length - 1) { // Removing: 最後に存在するもの
       newSelected = newSelected.concat(logIds.slice(0, -1))
-      removeCheck(expenditureLog)
+      removeCheck(incomeLog)
     } else if (selectedIndex > 0) { // Adding: 1番目以降にあるなら
       newSelected = newSelected.concat( // 3つの配列を連結させる
         logIds.slice(0, selectedIndex), // 最初〜取得した番号前まで取得
         logIds.slice(selectedIndex + 1), // 取得した番号+1番から最後まで取得
       )
 
-      removeCheck(expenditureLog)
+      removeCheck(incomeLog)
     }
   }
 
@@ -180,34 +182,34 @@ const ExpenditureLogsTable: React.FC = () => {
 
   const handleChangeDense = (event: React.ChangeEvent<HTMLInputElement>) => setDense(event.target.checked)
 
-  const isSelected = (row: ExpenditureLog) => {
+  const isSelected = (row: IncomeLog) => {
     return checkedLogs.map(log => { return log.id })
                       .indexOf(row.id) !== -1
   }
 
   const emptyRows = rowsPerPage - Math.min(rowsPerPage, rows.length - page * rowsPerPage)
 
-  const handleDeleteClick = (expenditureLog: ExpenditureLog) => {
-    if (window.confirm('Are you sure ?')) {
-      deleteExpenditureLog(expenditureLog)
-        .then((expenditureLog: ExpenditureLog) => {
-          dispatch(editExpenditureLog('DESTROY_EXPENDITURE_LOG', expenditureLog))
-          removeCheck(expenditureLog)
-        })
-        .catch(response => {
-          console.error(response)
-        })
-    }
+  const handleDeleteClick = (incomeLog: IncomeLog) => {
+    deleteIncomeLog(incomeLog)
+      .then((incomeLog: IncomeLog) => {
+        dispatch(editIncomeLog('DESTROY_INCOME_LOG', incomeLog))
+        removeCheck(incomeLog)
+        successMessage(succesmMessages.destroy)
+      })
+      .catch(response => {
+        console.error(response)
+      })
   }
 
   return (
     <React.Fragment>
-        <CreateExpenditureLogModal />
 
-        {expenditureLogs ? (
+        {incomeLogs ? (
           <div className={classes.root}>
+            <CreateIncomeLogModal />
+
             <Paper className={classes.paper}>
-              <EnhancedTableToolbar expenditureLogs={checkedLogs} numSelected={checkedLogs.length} setCheckedLogs={setCheckedLogs} />
+              <EnhancedTableToolbar incomeLogs={checkedLogs} numSelected={checkedLogs.length} setCheckedLogs={setCheckedLogs} />
               <TableContainer className={classes.container}>
                 <Table
                   stickyHeader
@@ -249,27 +251,23 @@ const ExpenditureLogsTable: React.FC = () => {
                                 onClick={(event) => handleCheckClick(event, row)}
                               />
                             </TableCell>
-                            <TableCell align='left'>
+                            <TableCell align="left">
                               {row.title}
                             </TableCell>
-                            <TableCell align='left'>
+                            <TableCell align="left">
                               {row.amount}
                             </TableCell>
-                            <TableCell align='left'>
+                            <TableCell align="left">
                               {row.content}
                             </TableCell>
-                            <TableCell align='left'>
-                              {moment(row.paidAt).format('YYYY-MM-DD')}
+                            <TableCell align="left">
+                              {moment(row.earnedAt).format('YYYY-MM-DD')}
                             </TableCell>
-                            <TableCell align='left'>
-                              <EdtingExpenditureLog expenditureLog={row} />
+                            <TableCell align="left">
+                              <EdtingIncomeLog incomeLog={row} />
                             </TableCell>
-                            <TableCell align='left'>
-                              <Tooltip title="Delete">
-                                <IconButton aria-label="delete" onClick={() => handleDeleteClick(row)}>
-                                  <DeleteForeverIcon />
-                                </IconButton>
-                              </Tooltip>
+                            <TableCell align="left">
+                              <DeleteAlert handleDeleteClick={handleDeleteClick} row={row} />
                             </TableCell>
                           </TableRow>
                         );
@@ -283,7 +281,7 @@ const ExpenditureLogsTable: React.FC = () => {
                 </Table>
               </TableContainer>
               <TablePagination
-                rowsPerPageOptions={[1, 7, expenditureLogs.length]}
+                rowsPerPageOptions={[1, 7, incomeLogs.length]}
                 component="div"
                 count={rows.length}
                 rowsPerPage={rowsPerPage}
@@ -297,8 +295,8 @@ const ExpenditureLogsTable: React.FC = () => {
               label="Dense padding"
             />
           </div>)
-        : null}
+        : <LoadingIcon />}
     </React.Fragment>
   );
 }
-export default ExpenditureLogsTable
+export default IncomeLogsTable
